fix: removed unneeded interface moved config paths
