+++ addon/ng2/blueprints/ng2/files/__path__/app/environments/environment.dev.ts
