+++ addon/ng2/blueprints/ng2/files/__path__/app/index.ts
@@ -1,2 +1,2 @@
-export * from './environment';
+export * from './environments/environment';
 export * from './app.component';
