+++ tests/e2e/e2e_workflow.spec.js
@@ -99,13 +99,13 @@ describe('Basic end-to-end Workflow', function () {
     expect(indexHtml).to.match(/app works!/);
   });
 
-  it('Supports production builds config file replacement', function() {
+  it('Supports build config file replacement', function() {
     this.timeout(420000);
 
     sh.exec(`${ngBin} build --dev`);
     var mainBundlePath = path.join(process.cwd(), 'dist', 'main.bundle.js');
     var mainBundleContent = fs.readFileSync(mainBundlePath, { encoding: 'utf8' });
-    // production: true minimized turns into production:!1
+
     expect(mainBundleContent).to.include('production: false');
   });
